rename builder files
