+++ packages/compiler/emitter-framework/index.ts
@@ -1,7 +1,7 @@
 export * from "./asset-emitter.js";
-export * from "./builders/array.js";
-export * from "./builders/object.js";
-export * from "./builders/string.js";
+export * from "./builders/array-builder.js";
+export * from "./builders/object-builder.js";
+export * from "./builders/string-builder.js";
 export * from "./placeholder.js";
 export * from "./type-emitter.js";
 export * from "./types.js";
